add search radius selector and improve useLocation with caching
